Factor out DomCrawlerNavigatorCallFactory::createElementCallArguments()
